Add flock core utility support
This patch adds flock support for glusterfs-coreutils. Unlike other
implemented core utils like cp, rm etc. flock will not be available
as a stand-alone command. Because locks are always associated with
file descriptors. So in order to keep those locks alive we need
active fds. Standalone commands like gfrm, gfls etc. immmediately
closes the connection after completing their corresponding jobs
which results in fd clean up and in turn to release locks for that
fd. A small bats test for basic flock funtionality testing is also
availble with this patch.

Signed-off-by: Anoop C S <anoopcs@redhat.com>