Fixes for exit code path from gluster shell on receving EOF
Inside gluster remote shell, quitting with Ctrl+D(EOF) forces users to
fallback into the same shell again. This fallback is due to a second
call to start the same shell as soon as we return from the first on
receving EOF. Apart from removing the second call, this change also
adds a missing newline after exiting the current shell.

Signed-off-by: Anoop C S <anoopcs@redhat.com>