Use GNU readline for interactive shell
Previously, we were using the basic getline() function
to retrieve input from the user for the interactive shell
gfcli(1). This was suboptimal since it makes adding extra
features such as shell history more tedious.

This changes introduces basic shell history courtesy of
the GNU history library. History is not persistent across
different sessions of the shell, only within a session.